Move AnsiEscapes & PRIV_CMD_EXEC to src/util.mjs
